# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas validadas",
                "description": "Obtiene empresas activas. Rate limit: 120 solicitudes por minuto por IP. Se recomienda filtrar por tipo cuando se integre desde clientes externos.",
                "operationId": "getEmpresas",
                "parameters": [
                    {
                        "name": "tipo_empresa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "contratacion-directa",
                                "est",
                                "outsourcing"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Registrar nueva empresa",
                "operationId": "createEmpresa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/empresas/{empresa}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "description": "Consulta de lectura con rate limit de 120 solicitudes por minuto por IP.",
                "operationId": "getEmpresa",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "operationId": "updateEmpresa",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Desactivar empresa",
                "description": "Desactiva el perfil sin eliminarlo de la base de datos. Rate limit: 30 solicitudes por minuto por IP.",
                "operationId": "deleteEmpresa",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa desactivada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/empresas/{empresa}/validar": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Validar empresa (solo administración)",
                "description": "Marca una empresa como validada. Rate limit: 30 solicitudes por minuto por IP.",
                "operationId": "validarEmpresa",
                "parameters": [
                    {
                        "name": "empresa",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa validada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmpresaValidationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "title": "Empresa",
                "description": "Modelo completo de empresa empleadora",
                "required": [
                    "nombre_empresa",
                    "rut_empresa",
                    "email",
                    "tipo_empresa",
                    "contacto_nombre",
                    "contacto_email"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "nombre_empresa": {
                        "type": "string",
                        "example": "TechCorp SpA"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "rrhh@techcorp.cl"
                    },
                    "logo_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología",
                        "nullable": true
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "enum": [
                            "contratacion-directa",
                            "est",
                            "outsourcing"
                        ]
                    },
                    "presentacion": {
                        "type": "string",
                        "nullable": true
                    },
                    "beneficios": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Seguro complementario",
                            "Trabajo remoto"
                        ],
                        "nullable": true
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@techcorp.cl"
                    },
                    "contacto_telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    },
                    "validado": {
                        "type": "boolean",
                        "example": false
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EmpresaInput": {
                "title": "Empresa Input",
                "description": "Datos para crear o actualizar una empresa",
                "required": [
                    "nombre_empresa",
                    "rut_empresa",
                    "email",
                    "tipo_empresa",
                    "contacto_nombre",
                    "contacto_email"
                ],
                "properties": {
                    "nombre_empresa": {
                        "type": "string",
                        "example": "TechCorp SpA"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "rrhh@techcorp.cl"
                    },
                    "logo_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología",
                        "nullable": true
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "enum": [
                            "contratacion-directa",
                            "est",
                            "outsourcing"
                        ]
                    },
                    "presentacion": {
                        "type": "string",
                        "nullable": true
                    },
                    "beneficios": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@techcorp.cl"
                    },
                    "contacto_telefono": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Respuesta estándar de error de la API",
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ActionMessage": {
                "title": "Action Message",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Persona desactivada exitosamente."
                    }
                },
                "type": "object"
            },
...
            "EmpresaValidationResult": {
                "title": "Empresa Validation Result",
                "required": [
                    "message",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Empresa validada exitosamente."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Empresa"
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "title": "Message Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActionMessage"
                    }
                },
                "type": "object"
            },
            "EmpresaListResponse": {
                "title": "Empresa List Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Empresa"
                        }
                    }
                },
                "type": "object"
            },
            "EmpresaResponse": {
                "title": "Empresa Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/Empresa"
                    }
                },
                "type": "object"
            },
            "EmpresaValidationResponse": {
                "title": "Empresa Validation Response",
                "required": [
                    "success",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/EmpresaValidationResult"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}